requirements.txt
fastapi[all]
uvicorn
ruff
pylint
sqlalchemy
coverage
pytest
pytest-cov
jwt
python-dotenv
pymysql
passlib[argon2]
python-jose[cryptography]
requests